Changed Seed to bigint

+++ db/migrate/20140415221454_add_seed_to_users.rb
@@ -1,5 +1,5 @@
 class AddSeedToUsers < ActiveRecord::Migration
   def change
-  	add_column :users, :seed, :integer
+  	add_column :users, :seed, :bigint
   end
 end
